Previous push refactored mouse and key codes to be platform agnostic, current push refactors IsoEngine's input class entirely to be platform agnostic, fixed some warnings for implicit type conversions.

+++ IsoEngine/src/IsoEngine/Debug/Instrumentor.h
@@ -154,7 +154,7 @@ namespace IE
 			auto HR_StartTime = FP_MICROSECONDS{ m_StartTimepoint.time_since_epoch() };
 			auto HR_EndTime = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch() - std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch();
 
-			uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+			uint32_t threadID = static_cast<uint32_t>(std::hash<std::thread::id>{}(std::this_thread::get_id()));
 			Instrumentor::Get().WriteProfile({ m_Name, HR_StartTime, HR_EndTime, threadID });
 
 			m_Stopped = true;
