+++ IsoEngine/src/IsoEngine/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -81,7 +81,7 @@ namespace IE
 		}
 
 		m_VertexBuffers.push_back(vertexBuffer);
-		m_VertexBufferIndexOffset += layout.GetElements().size();
+		m_VertexBufferIndexOffset += static_cast<uint32_t>(layout.GetElements().size());
 	}
 
 	void OpenGLVertexArray::SetIndexBuffer(const Ref<IndexBuffer>& indexBuffer)
